components/function-runtimes/python39/kubeless/requirements.txt
requests
bottle==0.12.19
cherrypy==8.9.1
wsgi-request-logger==0.4.6
prometheus_client==0.8.0
opentelemetry-api==1.9.1
opentelemetry-sdk==1.9.1
opentelemetry-exporter-jaeger==1.9.1
opentelemetry-propagator-b3==1.9.1
opentelemetry-instrumentation-requests==0.28b1